Feature/connection professional services

+++ back-end/ms-professionals/src/main/java/petcareveterinary/employees/model/Professional.java
@@ -3,6 +3,7 @@ package petcareveterinary.employees.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+
 @Data
 @Entity
 @Table(name="Professionals")
@@ -13,5 +14,5 @@ public class Professional {
     @Column
     private String licenseNumber;
     @Column(name = "user_id")
-    private int userId;
+    private Long userId;
 }
